Extract operator deserialization into a new module
The `model` module was becoming rather large. Reduce its size by extracting the
`OpRegistry` type and associated code for deserializing operators into a
separate module.
